messeges

+++ bank/app/Http/Controllers/ClientController.php
@@ -8,6 +8,12 @@ use Illuminate\Support\Facades\Validator;
 
 class ClientController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     
     public function index()
     {
@@ -30,7 +36,7 @@ class ClientController extends Controller
         $validator = Validator::make($request->all(), [
             'name' => 'required|alpha:ascii|min:3',
             'surname' => 'required|alpha:ascii|min:3',
-            'idPerson' => 'required|numeric|min_digits:11|max_digits:11'
+            'idPerson' => 'required|numeric|unique:clients|min_digits:11|max_digits:11'
         ],
         [
             'name.min' => 'Name is too short - should be at least 3 characters',
@@ -46,7 +52,6 @@ class ClientController extends Controller
         }
 
 
-
         $client = new Client;
         $client->name = $request->name;
         $client->surname = $request->surname;
@@ -55,7 +60,8 @@ class ClientController extends Controller
         $client->amount = $request->amount;
         $client->save();
         return redirect()
-        ->route('clients-index');
+        ->route('clients-index')
+        ->with('ok', 'New client was created');
 
     }
 
@@ -85,27 +91,42 @@ class ClientController extends Controller
     
     public function update(Request $request, Client $client)
     {
+
         $amountPlus = $client->amount + $request->amount;
         $client->amount = $amountPlus;
         $client->save();
         return redirect()
-            ->route('clients-index');
+            ->route('clients-index')
+            ->with('info', 'Amount was added');
     }
 
     public function updateDeduct(Request $request, Client $client)
     {
         $amountMinus = $client->amount - $request->amount;
+        if ($client->amount < $request->amount) {
+            return redirect()
+                ->route('clients-editDeduct', $client)
+                ->with('info','You want to deduct too much!');
+        }
         $client->amount = $amountMinus;
         $client->save();
         return redirect()
-            ->route('clients-index');
+            ->route('clients-index')
+            ->with('info', 'Amount was deducted');
     }
 
    
     public function destroy(Client $client)
     {
+        if ($client->amount > 0) {
+            return redirect()
+                ->route('clients-index')
+                ->with('info','Client account not empty!');
+        }
+
         $client->delete();
         return redirect()
-        ->route('clients-index');
+        ->route('clients-index')
+        ->with('info', 'The client was deleted!');
     }
 }
